requirements.txt
# Requirements file for MediPlant

# Flask Framework
Flask==2.3.3
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-Caching==2.1.0
Flask-Limiter==3.5.0

# Database
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
cryptography==41.0.4

# Forms and Validation
WTForms==3.0.1

# Password Hashing
bcrypt==4.0.1
Werkzeug==2.3.7

# Image Processing
Pillow==10.0.1

# Date and Time
python-dateutil==2.8.2

# Environment Variables
python-dotenv==1.0.0

# Payment Gateways
razorpay==1.4.1

# AWS SDK (for S3 file storage)
boto3==1.28.62

# Redis (for caching and sessions)
redis==5.0.0

# HTTP Requests
requests==2.31.0

# JSON handling
orjson==3.9.7

# Testing
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0

# Development Tools
flask-shell-ipython==1.4.0

# Production Server
gunicorn==21.2.0

# Monitoring and Logging
sentry-sdk[flask]==1.32.0

# Security
Flask-Talisman==1.1.0

# API Documentation
flask-restx==1.2.0

# Background Tasks
celery==5.3.2

# Email Templates
flask-mailman==0.3.0

# File Upload
flask-uploads==0.2.1

# Admin Interface
Flask-Admin==1.6.1

# Social Authentication (optional)
Flask-OAuthlib==0.9.6

# Rate Limiting
slowapi==0.1.8

# Data Validation
marshmallow==3.20.1
flask-marshmallow==0.15.0

# Timezone Support
pytz==2023.3

# UUID Generation
uuid==1.30

# CSV Processing
pandas==2.1.1

# Excel File Processing
openpyxl==3.1.2

# QR Code Generation
qrcode==7.4.2

# Barcode Generation
python-barcode==0.14.0

# PDF Generation
reportlab==4.0.4

# Web Scraping (for price comparison)
beautifulsoup4==4.12.2

# HTTP Client
httpx==0.25.0

# URL Parsing
urllib3==2.0.5

# Configuration Management
dynaconf==3.2.3

# Logging
structlog==23.1.0

# Task Queue
rq==1.15.1

# Background Job Scheduler
APScheduler==3.10.4

# Health Checks
flask-healthz==0.0.3

# CORS Support
Flask-CORS==4.0.0

# API Throttling
flask-limiter==3.5.0

# Session Management
Flask-Session==0.5.0

# CSRF Protection
Flask-SeaSurf==1.1.1

# Content Security Policy
flask-csp==1.3.0

# Development and Debugging
flask-debugtoolbar==0.13.1
